responded to comments in luca branch
